Optimization for retina smartphones
The srcset technique misses an important use case for larger images:
if the user has a retina smartphone, it'll download the high-resolution
image even when the standard version would suffice.
We want to optimize larger images based on their final rendered
dimensions, not just the device's screen resolution.

Our low-resolution photo is 1000 pixels wide, which means that
(with this commit) 2x retina devices can use it as
long as their screen is less than 500 pixels wide.

Now we can serve a 115KB image to mobile devices instead of forcing
them to use the high-res 445KB image.

To test this on Chrome, we have to avoid the local browser cache by
opening a new incognito window, then avoid loading photo-big.jpg by
making the window very narrow before loading the page.