Art direction using <picture>
We can optimize layouts by sending completely different images to the
user depending on their device. Previously, instead, we optimized the
same image for different devices.

This level of control will make the designer very happy, but the trade
off is that it doesn't let the browser automatically pick the optimal
image. This means that the (previously implemented) retina optimization
is lost: as long as the screen width is 401 pixels or greater, the
browser will always use the high-resolution, wide-cropped image.

While it is possible to combine the best of both worlds, it gets
complicated real quick. IIH's recommendation is to stick to the 1x and
2x version of srcset for images less than 600 pixels wide, use the
srcset plus sizes method from the previous section for bigger photos,
and reserve <picture> for when trying to do something real fancy with
the designer guy/gal.